requirements.txt
fastapi
uvicorn
scikit-learn
pandas
numpy
joblib
python-multipart
python-dotenv
